fewer complaints when the mail log doesn't exist.
git-svn-id: svn+ssh://svn.bestpractical.com/svn/bps-public/rt/branches/3.999-DANGEROUS@10166 e417ac7c-1bcc-0310-8ffa-8f5827389a85

+++ lib/RT/Test.pm
@@ -361,13 +361,15 @@ sub file_content {
     $path = File::Spec->catfile( @$path ) if ref $path;
 
     diag "reading content of '$path'" if $ENV{'TEST_VERBOSE'};
-
+    my $content = '';
+    if (-f $path) {
     open my $fh, "<:raw", $path
         or do { warn "couldn't open file '$path': $!"; return '' };
-    my $content = do { local $/; <$fh> };
+    $content = do { local $/; <$fh> };
     close $fh;
-
+    
     unlink $path if $args{'unlink'};
+    }
 
     return $content;
 }
